Primer boletín de colecciones completado

+++ PROG2526-Unidad03.3/src/boletin01/Ejercicio02.java
@@ -6,31 +6,28 @@ import java.util.Random;
 public class Ejercicio02 {
 
 	public static void main(String[] args) {
-
-		/*
-		 * -----------------------------------------------------------------------------
-		 * Realiza un programa que introduzca valores aleatorios (entre 0 y 100) en un
-		 * ArrayList y que luego calcule la suma, la media, el máximo y el mínimo de
-		 * esos números. El tamaño de la lista también será aleatorio y podrá oscilar
-		 * entre 5 y 10 elementos ambos inclusive.
-		 * -----------------------------------------------------------------------------
-		 */
-
 		Random rnd = new Random();
 
+		// CREACIÓN DE LISTA
 		ArrayList<Integer> valoresAleatorios = new ArrayList<>();
 
 		int tamano = rnd.nextInt(5, 11);
 		int numero;
 		int suma = 0;
-		int media = 0;
+		double media = 0;
+
+		// VARIABLES AUXILIARES PARA CALCULAR EL MÍNIMO Y MÁXIMO
 		int maximo = Integer.MIN_VALUE;
 		int minimo = Integer.MAX_VALUE;
 
+		// ASIGNACIÓN DE VALORES ALEATORIOS EN LA LISTA
 		for (int i = 0; i < tamano; i++) {
 			numero = rnd.nextInt(0, 101);
 			valoresAleatorios.add(numero);
+			// SUMA TOTAL
 			suma += numero;
+
+			// COMPROBACIÓN DE MÍNIMO Y MÁXIMO
 			if (numero > maximo) {
 				maximo = numero;
 			}
@@ -39,8 +36,9 @@ public class Ejercicio02 {
 			}
 		}
 
-		media = suma / tamano;
+		media = (double) suma / tamano;
 
+		// IMPRIME RESULTADOS
 		System.out.println(valoresAleatorios);
 		System.out.println("Suma total: " + suma);
 		System.out.println("Media: " + media);
